Removed DS_Store files and updated scripts

+++ active_learning/model/data_generation_wrapper.py
@@ -189,15 +189,15 @@ def compileCASMOutput(rnd, casm_version, len,temp=''):
 def loadCASMOutput(rnd,dim,outputFolder,singleRnd=False):
 
     if singleRnd:
-        kappa = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float32)[:,:dim]
-        eta = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float32)[:,dim:2*dim]
-        mu = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float32)[:,-dim:]
-        T = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float32)[:,-dim-1:-dim]
+        kappa = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float64)[:,:dim]
+        eta = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float64)[:,dim:2*dim]
+        mu = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float64)[:,-dim:]
+        T = np.genfromtxt(outputFolder + 'data/data_sampled/results'+str(rnd)+'.txt',dtype=np.float64)[:,-dim-1:-dim]
     else:
-        kappa = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float32)[:,:dim]
-        eta = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float32)[:,dim:2*dim]
-        mu = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float32)[:,-dim:]
-        T = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float32)[:,-dim-1:-dim]
+        kappa = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float64)[:,:dim]
+        eta = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float64)[:,dim:2*dim]
+        mu = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float64)[:,-dim:]
+        T = np.genfromtxt(outputFolder + 'data/data_sampled/allResults'+str(rnd)+'.txt',dtype=np.float64)[:,-dim-1:-dim]
 
     return kappa, eta, mu, T
